chore: refactor main container
