+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@uniformdev/uniformconf",
-  "version": "17.6.0",
+  "version": "18.8.0",
   "private": true,
   "license": "Apache-2.0",
   "scripts": {
